refactored mvv_lva
